// src/Juniper.TypeScript/@juniper-lib/threejs/ScreenMode.d.ts
export declare enum ScreenMode {
    None = "None",
    Fullscreen = "Fullscreen",
    VR = "VR",
    AR = "AR",
    Anaglyph = "Anaglyph",
    FullscreenAnaglyph = "FullscreenAnaglyph"
}
export type ScreenModeSelection = Exclude<ScreenMode, ScreenMode.FullscreenAnaglyph>;
//# sourceMappingURL=ScreenMode.d.ts.map